msm: kgsl: Return the correct error from kgsl_active_count_wait()
kgsl_active_count_wait() is very badly crafted. If the active count
is already under the threshold (as it often is - a successful condition)
the conditional at the end of the function returns -ETIMEDOUT instead
of 0. Return the proper error value so the caller can make a wise
decision based on the result.

Change-Id: Ic0dedbadb948820cf64e3131f895fa3d15247cac
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>